requirements.txt
fastapi==0.115.2
uvicorn==0.26.0
langchain==0.3.26
langchain-anthropic==0.3.17
langchain-core==0.3.69
langchain-neo4j==0.4.0
neo4j==5.28.1
python-dotenv==1.0.1
pydantic==2.11.7
numpy==1.26.2
langsmith==0.4.8
pandas==2.1.3
starlette==0.40.0
python-multipart==0.0.18
